Added IAM-VideoServiceWithDynamoDBBeanstalk
Tested & working & deployed to AWS

In this version of the VideoSvc application, we are using DynamoDB to
store data.
See the Video class for the annotation changes that this requires.

See the Application class for the configuration of your Amazon AWS
credentials.
In this case, using IAM and the method
InstanceProfileCredentialsProvider()
to dynamically provide access to DynamoDB using a EC2 instance profile
and a
IAM role with access policies.

An ApplicationServletInitializer class has been added to boostrap the
application
when it is deployed to a stand-alone Tomact instance in Amazon Elastic
Beanstalk.